+++ src/rust/bitbox02-rust/src/hww/api.rs
@@ -25,6 +25,7 @@ mod device_info;
 mod electrum;
 mod error;
 mod reset;
+mod restore;
 mod rootfingerprint;
 mod sdcard;
 mod set_device_name;
@@ -122,6 +123,7 @@ async fn process_api(request: &Request) -> Option<Result<Response, Error>> {
         Request::CheckBackup(ref request) => Some(backup::check(request).await),
         Request::CreateBackup(ref request) => Some(backup::create(request).await),
         Request::ShowMnemonic(_) => Some(show_mnemonic::process().await),
+        Request::RestoreFromMnemonic(ref request) => Some(restore::from_mnemonic(request).await),
         Request::ElectrumEncryptionKey(ref request) => Some(electrum::process(request).await),
 
         #[cfg(feature = "app-ethereum")]