+++ src/rust/bitbox02/src/keystore.rs
@@ -226,3 +226,17 @@ pub fn root_fingerprint() -> Result<[u8; 4], ()> {
         false => Err(()),
     }
 }
+
+pub fn encrypt_and_store_seed(seed: &[u8], password: &str) -> Result<(), ()> {
+    let password = zeroize::Zeroizing::new(crate::util::str_to_cstr_vec(password)?);
+    match unsafe {
+        bitbox02_sys::keystore_encrypt_and_store_seed(
+            seed.as_ptr(),
+            seed.len() as _,
+            password.as_ptr(),
+        )
+    } {
+        true => Ok(()),
+        false => Err(()),
+    }
+}