+++ test/unit-test/test_util.c
@@ -57,13 +57,13 @@ static void test_util_format_datetime(void** state)
     assert_string_equal(out, "Mon 2020-09-28");
 
     util_format_datetime(1601281809, 0, false, out, sizeof(out));
-    assert_string_equal(out, "Mon 2020-09-28\n08:30:09");
+    assert_string_equal(out, "Mon 2020-09-28\n08:30");
 
     util_format_datetime(1601281809, 18000, false, out, sizeof(out));
-    assert_string_equal(out, "Mon 2020-09-28\n13:30:09");
+    assert_string_equal(out, "Mon 2020-09-28\n13:30");
 
     util_format_datetime(1601281809, -32400, false, out, sizeof(out));
-    assert_string_equal(out, "Sun 2020-09-27\n23:30:09");
+    assert_string_equal(out, "Sun 2020-09-27\n23:30");
 }
 
 int main(void)