+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [Unreleased]
--
+- RestoreFrommnemonic: ported to Rust. Will now return UserAbortError on user abort instead of GenericError.
 
 ## 9.4.0 [released 2021-01-20]
 - ETHPubRequest api call now fails if a an invalid contract address is provided also if `display` is